Moved file and process service into dedicated service package under the util package
