+++ src/main/java/gr/aegean/icsd/icarus/util/services/FileService.java
@@ -1,4 +1,4 @@
-package gr.aegean.icsd.icarus.util;
+package gr.aegean.icsd.icarus.util.services;
 
 import gr.aegean.icsd.icarus.util.exceptions.async.TestExecutionFailedException;
 import jakarta.validation.constraints.NotBlank;
