+++ src/main/java/gr/aegean/icsd/icarus/util/terraform/StackDeployer.java
@@ -6,8 +6,8 @@ import gr.aegean.icsd.icarus.provideraccount.GcpAccount;
 import gr.aegean.icsd.icarus.provideraccount.ProviderAccount;
 import gr.aegean.icsd.icarus.resourceconfiguration.ResourceConfiguration;
 import gr.aegean.icsd.icarus.test.Test;
-import gr.aegean.icsd.icarus.util.FileService;
-import gr.aegean.icsd.icarus.util.ProcessService;
+import gr.aegean.icsd.icarus.util.services.FileService;
+import gr.aegean.icsd.icarus.util.services.ProcessService;
 import gr.aegean.icsd.icarus.util.aws.AwsRegion;
 import gr.aegean.icsd.icarus.util.aws.LambdaRuntime;
 import gr.aegean.icsd.icarus.util.enums.Platform;
